+++ .babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    "@babel/react"
+  ],
+  "plugins": [
+    "react-hot-loader/babel"
+  ]
+}
