+++ legacy2/.gitignore
@@ -1,3 +1,3 @@
 node_modules
-build
+# build
 secrets.*.js
