+++ legacy2/package.json
@@ -1,0 +1,26 @@
+{
+	"name": "chrome-extension-webpack",
+	"version": "0.0.1",
+	"description": "A boilerplate to chrome extension with webpack",
+	"scripts": {
+		"build": "node utils/build.js",
+		"start": "node utils/webserver.js"
+	},
+	"devDependencies": {
+		"clean-webpack-plugin": "3.0.0",
+		"copy-webpack-plugin": "5.0.5",
+		"css-loader": "3.2.0",
+		"file-loader": "4.3.0",
+		"fs-extra": "8.1.0",
+		"html-loader": "0.5.5",
+		"html-webpack-plugin": "3.2.0",
+		"style-loader": "1.0.0",
+		"webpack": "4.41.2",
+		"webpack-cli": "3.3.10",
+		"webpack-dev-server": "3.9.0",
+		"write-file-webpack-plugin": "4.5.1"
+	},
+	"dependencies": {
+		"jikanjs": "^0.7.0"
+	}
+}
