+++ legacy2/src/background.html
@@ -4,5 +4,5 @@
 		<meta charset="UTF-8" />
 		<title></title>
 	</head>
-	<body><script type="text/javascript" src="background.bundle.js"></script></body>
+	<body></body>
 </html>
