+++ legacy2/src/manifest.json
