+++ legacy2/src/options.html
@@ -10,7 +10,7 @@
 				<h1 class="logo-title">
 					<img
 						class="logo"
-						src="aniTracker32.png"
+						src="./img/aniTracker32.png"
 						alt="aniTrack Logo"
 					/>
 					aniTracker
@@ -19,7 +19,7 @@
 					<a class="active" href="#">
 						<img
 							class="menu-icon home-icon"
-							src="home.svg"
+							src="./img/icons/home.svg"
 							alt="Home Icon"
 						/>
 						Home
@@ -27,7 +27,7 @@
 					<a href="#">
 						<img
 							class="menu-icon sync-icon"
-							src="copy.svg"
+							src="./img/icons/copy.svg"
 							alt="Copy Icon"
 						/>
 						Sync With MyAnimeList
@@ -35,7 +35,7 @@
 					<a href="#">
 						<img
 							class="menu-icon heart-icon"
-							src="heart.svg"
+							src="./img/icons/heart.svg"
 							alt="Heart Icon"
 						/>
 						Support
@@ -43,7 +43,7 @@
 					<a href="#">
 						<img
 							class="menu-icon settings-icon"
-							src="settings.svg"
+							src="./img/icons/settings.svg"
 							alt="Cog Icon"
 						/>
 						Settings
@@ -59,7 +59,7 @@
 							<input class="list-searchbar" type="text" />
 							<img
 								class="search-icon"
-								src="search.svg"
+								src="./img/icons/search.svg"
 								alt="search icon"
 							/>
 						</div>
@@ -86,13 +86,13 @@
 									<a id="episode-sort-most" href="#"
 										><img
 											class="up-arrow"
-											src="chevron-up.svg"
+											src="./img/icons/chevron-up.svg"
 											alt="Up arrow"
 									/></a>
 									<a id="episode-sort-least" href="#"
 										><img
 											class="down-arrow"
-											src="chevron-down.svg"
+											src="./img/icons/chevron-down.svg"
 											alt="Down arrow"
 									/></a>
 								</div>
@@ -106,13 +106,13 @@
 									<a id="sort-completion-most" href="#"
 										><img
 											class="up-arrow"
-											src="chevron-up.svg"
+											src="./img/icons/chevron-up.svg"
 											alt="Up arrow"
 									/></a>
 									<a id="sort-completion-least" href="#"
 										><img
 											class="down-arrow"
-											src="chevron-down.svg"
+											src="./img/icons/chevron-down.svg"
 											alt="Down arrow"
 									/></a>
 								</div>
@@ -194,5 +194,5 @@
 				</div> -->
 			</section>
 		</div>
-	<script type="text/javascript" src="options.bundle.js"></script></body>
+	</body>
 </html>
