+++ legacy2/src/popup.html
@@ -40,5 +40,5 @@
 				<textarea name="" id="" cols="30" rows="10"></textarea>
 			</div>
 		</div>
-	<script type="text/javascript" src="popup.bundle.js"></script></body>
+	</body>
 </html>
