+++ legacy2/utils/build.js
@@ -1,0 +1,9 @@
+var webpack = require("webpack"),
+    config = require("../webpack.config");
+
+delete config.chromeExtensionBoilerplate;
+
+webpack(
+  config,
+  function (err) { if (err) throw err; }
+);
