+++ legacy2/utils/env.js
@@ -1,0 +1,5 @@
+// tiny wrapper with default env vars
+module.exports = {
+  NODE_ENV: (process.env.NODE_ENV || "development"),
+  PORT: (process.env.PORT || 3000)
+};
