+++ package.json
@@ -1,26 +1,31 @@
 {
-	"name": "chrome-extension-webpack",
-	"version": "0.0.1",
-	"description": "A boilerplate to chrome extension with webpack",
-	"scripts": {
-		"build": "node utils/build.js",
-		"start": "node utils/webserver.js"
-	},
-	"devDependencies": {
-		"clean-webpack-plugin": "3.0.0",
-		"copy-webpack-plugin": "5.0.5",
-		"css-loader": "3.2.0",
-		"file-loader": "4.3.0",
-		"fs-extra": "8.1.0",
-		"html-loader": "0.5.5",
-		"html-webpack-plugin": "3.2.0",
-		"style-loader": "1.0.0",
-		"webpack": "4.41.2",
-		"webpack-cli": "3.3.10",
-		"webpack-dev-server": "3.9.0",
-		"write-file-webpack-plugin": "4.5.1"
-	},
-	"dependencies": {
-		"jikanjs": "^0.7.0"
-	}
+  "name": "chrome-extension-webpack",
+  "version": "0.0.1",
+  "description": "A boilerplate to chrome extension with webpack",
+  "scripts": {
+    "build": "node utils/build.js",
+    "start": "node utils/webserver.js"
+  },
+  "dependencies": {
+    "react": "16.12.0",
+    "react-dom": "16.12.0"
+  },
+  "devDependencies": {
+    "@babel/core": "7.7.4",
+    "@babel/preset-react": "7.7.4",
+    "babel-loader": "8.0.6",
+    "react-hot-loader": "4.12.18",
+    "clean-webpack-plugin": "3.0.0",
+    "copy-webpack-plugin": "5.0.5",
+    "css-loader": "3.2.0",
+    "file-loader": "4.3.0",
+    "fs-extra": "8.1.0",
+    "html-loader": "0.5.5",
+    "html-webpack-plugin": "3.2.0",
+    "style-loader": "1.0.0",
+    "webpack": "4.41.2",
+    "webpack-dev-server": "3.9.0",
+    "write-file-webpack-plugin": "4.5.1",
+    "webpack-cli": "3.3.10"
+  }
 }
