+++ src/background.html
@@ -1,8 +1,9 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<title></title>
-	</head>
-	<body></body>
+<head>
+  <meta charset="UTF-8">
+  <title></title>
+</head>
+<body>
+</body>
 </html>
