--- src/css/popup.css
@@ -1,190 +1,0 @@
-html {
-	margin: 0;
-	padding: 0;
-	border: 0;
-	font-size: 100%;
-	font: inherit;
-	vertical-align: baseline;
-}
-
-body {
-	color: #404040;
-	font-family: "Lucida Grande", Verdana, sans-serif;
-	font-size: 13px;
-	font-weight: normal;
-	line-height: 20px;
-	background: #3b3f46;
-	margin: 0 auto;
-	padding: 20px;
-	width: 200px;
-	background-color: false;
-	background-image: -webkit-gradient(
-		radial,
-		center center,
-		0,
-		center center,
-		460,
-		color-stop(0%, rgba(255, 255, 255, 0.1)),
-		color-stop(100%, transparent)
-	);
-	background-image: -webkit-radial-gradient(
-		center,
-		cover,
-		rgba(255, 255, 255, 0.1),
-		transparent
-	);
-	background-image: -moz-radial-gradient(
-		center,
-		cover,
-		rgba(255, 255, 255, 0.1),
-		transparent
-	);
-	background-image: -ms-radial-gradient(
-		center,
-		cover,
-		rgba(255, 255, 255, 0.1),
-		transparent
-	);
-	background-image: -o-radial-gradient(
-		center,
-		cover,
-		rgba(255, 255, 255, 0.1),
-		transparent
-	);
-	background-image: radial-gradient(
-		center,
-		cover,
-		rgba(255, 255, 255, 0.1),
-		transparent
-	);
-}
-
-.container {
-	text-align: center;
-}
-
-/* .no-track {
-				position: absolute;
-				top: 0;
-				left: 0;
-				bottom: 0;
-				right: 0;
-				background-color: rgba(0, 0, 0, 0.7);
-				z-index: 3;
-			}
-
-			.no-track-off {
-				display: none;
-			}
-
-			.no-track-text {
-				color: white;
-				font-size: 20px;
-				position: absolute;
-				left: 0;
-				right: 0;
-				top: 50%;
-				transform: translateY(-50%);
-				text-shadow: 0 2px 4px rgba(28, 206, 43, 0.7);
-			} */
-
-.track-status {
-	font-size: 18px;
-	color: #f08080;
-}
-
-.track-status-active {
-	color: #65bd63;
-}
-
-.switch {
-	position: relative;
-	display: inline-block;
-	height: 26px;
-	width: 120px;
-	background: rgba(0, 0, 0, 0.25);
-	border-radius: 3px;
-	-webkit-box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.3),
-		0 1px rgba(255, 255, 255, 0.1);
-	box-shadow: inset 0 1px 3px rgba(0, 0, 0, 0.3),
-		0 1px rgba(255, 255, 255, 0.1);
-	vertical-align: middle;
-}
-
-.switch-name {
-	color: white;
-	font-size: 18px;
-	height: 26px;
-	vertical-align: middle;
-	margin-right: 10px;
-}
-
-.switch-label {
-	position: relative;
-	z-index: 2;
-	float: left;
-	width: 58px;
-	line-height: 26px;
-	font-size: 11px;
-	color: rgba(255, 255, 255, 0.35);
-	text-align: center;
-	text-shadow: 0 1px 1px rgba(0, 0, 0, 0.45);
-	cursor: pointer;
-}
-
-.switch-label:active {
-	font-weight: bold;
-}
-
-.switch-label-off {
-	padding-left: 2px;
-}
-
-.switch-label-on {
-	padding-right: 2px;
-}
-
-.switch-input {
-	display: none;
-}
-
-.switch-input:checked + .switch-label {
-	font-weight: bold;
-	color: rgba(0, 0, 0, 0.65);
-	text-shadow: 0 1px rgba(255, 255, 255, 0.25);
-	-webkit-transition: 0.15s ease-out;
-	-moz-transition: 0.15s ease-out;
-	-ms-transition: 0.15s ease-out;
-	-o-transition: 0.15s ease-out;
-	transition: 0.15s ease-out;
-	-webkit-transition-property: color, text-shadow;
-	-moz-transition-property: color, text-shadow;
-	-ms-transition-property: color, text-shadow;
-	-o-transition-property: color, text-shadow;
-	transition-property: color, text-shadow;
-}
-
-.switch-input:checked + .switch-label-on ~ .switch-selection {
-	left: 60px;
-	/* Note: left: 50%; doesn't transition in WebKit */
-}
-
-.switch-selection {
-	position: absolute;
-	z-index: 1;
-	top: 2px;
-	left: 2px;
-	display: block;
-	width: 58px;
-	height: 22px;
-	border-radius: 3px;
-	background-color: #65bd63;
-	-webkit-box-shadow: inset 0 1px rgba(255, 255, 255, 0.5),
-		0 0 2px rgba(0, 0, 0, 0.2);
-	box-shadow: inset 0 1px rgba(255, 255, 255, 0.5), 0 0 2px rgba(0, 0, 0, 0.2);
-	-webkit-transition: left 0.15s ease-out;
-	-moz-transition: left 0.15s ease-out;
-	-ms-transition: left 0.15s ease-out;
-	-o-transition: left 0.15s ease-out;
-	transition: left 0.15s ease-out;
-}
