+++ src/manifest.json
@@ -1,26 +1,16 @@
 {
-	"name": "Weeb Track",
-	"version": "0.0.1",
-	"description": "Keeps track of what episode you are on",
-	"author": "John Duy Dinh",
-	"permissions": [
-		"declarativeContent",
-		"tabs",
-		"storage",
-		"activeTab",
-		"unlimitedStorage"
-	],
-	"options_page": "options.html",
-	"background": {
-		"page": "background.html"
-	},
-	"browser_action": {
-		"default_popup": "popup.html",
-		"default_icon": "aniTracker32.png"
-	},
-	"icons": {
-		"48": "aniTracker32.png"
-	},
-	"manifest_version": 2,
-	"content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
+  "name": "Chrome Extension Webpack",
+  "options_page": "options.html",
+  "background": {
+    "page": "background.html"
+  },
+  "browser_action": {
+    "default_popup": "popup.html",
+    "default_icon": "icon-34.png"
+  },
+  "icons": {
+    "128": "icon-128.png"
+  },
+  "manifest_version": 2,
+  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
 }
