+++ src/popup.html
@@ -1,44 +1,10 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta charset="UTF-8" />
-		<title></title>
-	</head>
-	<body>
-		<div class="container">
-			<!-- <div class="no-track">
-				<span class="no-track-text">
-					Can't Track This Page
-				</span>
-			</div> -->
-			<h1 class="track-status">Page Not Compatible</h1>
-			<span class="switch-name" style="color: white">Track: </span>
-			<div class="switch">
-				<input
-					type="radio"
-					class="switch-input"
-					name="view"
-					id="on"
-					checked
-				/>
-				<label for="on" class="switch-label switch-label-off">ON</label>
-				<input type="radio" class="switch-input" name="view" id="off" />
-				<label for="off" class="switch-label switch-label-on"
-					>OFF</label
-				>
-				<span class="switch-selection"></span>
-				<button
-					id="list"
-					style="padding: 20px; background-color: #172b4d; color: white; font-size: 18px; border: none;"
-				>
-					View List
-				</button>
-				<button id="reset">Reset List</button>
-				<a style="font-size: 18px; color: white;" id="export"
-					>Export Save Data</a
-				>
-				<textarea name="" id="" cols="30" rows="10"></textarea>
-			</div>
-		</div>
-	</body>
+<head>
+  <meta charset="UTF-8">
+  <title></title>
+</head>
+<body>
+  <div id="app-container"></div>
+</body>
 </html>
